requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
sentence-transformers==2.6.0
huggingface_hub==0.22.2
transformers==4.37.2
accelerate==0.29.3
torch==2.2.0  # Compatible with Python 3.9
newspaper3k
fastapi==0.110.0
uvicorn==0.32.0
pydantic==2.7.0
